the strictening, part 1: mandatory annotations, no exceptions
One day a glorious day will dawn when Python will be a strongly, statically typed language, and the last vestiges of hackery and runtime error checking will be relegated to the dark past. That day is coming, and we can all help bring it closer! Of course we'll always be able to look away if you turn that off for the really tricky bits.